refactor: remove scannable_errors.py
Having error handling in a separate file would cause issue with the next
commit because it makes `handle_scan_chunk_error()` depend on `File`.

Move `handle_scan_chunk_error()` to `scannable.py` and
`handle_scan_error()` to `iac/scan.py`.

Adjust tests accordingly.
