+++ ggshield/cmd/iac/scan.py
@@ -1,12 +1,15 @@
+import logging
 from pathlib import Path
 from typing import Any, Optional, Sequence, Type
 
 import click
+from pygitguardian.models import Detail
 
 from ggshield.core.client import create_client_from_config
 from ggshield.core.config import Config
 from ggshield.core.extra_headers import get_headers
 from ggshield.core.filter import init_exclusion_regexes
+from ggshield.core.text_utils import display_error
 from ggshield.core.utils import ScanContext
 from ggshield.iac.filter import get_iac_files_from_paths
 from ggshield.iac.models import IaCScanResult
@@ -16,7 +19,9 @@ from ggshield.output import OutputHandler
 from ggshield.output.json.iac_json_output_handler import IaCJSONOutputHandler
 from ggshield.output.text.iac_text_output_handler import IaCTextOutputHandler
 from ggshield.scan import ScanCollection
-from ggshield.scan.scannable_errors import handle_scan_error
+
+
+logger = logging.getLogger(__name__)
 
 
 def validate_exclude(_ctx: Any, _param: Any, value: Sequence[str]) -> Sequence[str]:
@@ -162,3 +167,11 @@ def iac_scan(ctx: click.Context, directory: Path) -> Optional[IaCScanResult]:
         handle_scan_error(scan)
         return None
     return scan
+
+
+def handle_scan_error(detail: Detail) -> None:
+    logger.error("status_code=%d detail=%s", detail.status_code, detail.detail)
+    if detail.status_code == 401:
+        raise click.UsageError(detail.detail)
+    display_error("\nError scanning.")
+    display_error(str(detail))
